+++ catmull_rom_spline.hpp
@@ -11,18 +11,18 @@
 template<size_t N, size_t n_points, size_t dimensions>
 class CatmullRomSpline {
 public:
-  explicit CatmullRomSpline(const std::array<Point<dimensions>, N>& p, const float tao) : p(p), tao(tao) {
-    const float interval = (1.)/n_points;
-    std::array<std::array<float, 4>,  n_points+1> coefficients;
-    const std::array<std::array<float,4>, 4> tao_matrix {
-      std::array<float,4>{0., 1., 0., 0.},
-      std::array<float,4>{-tao, 0., tao, 0.},
-      std::array<float,4>{2.*tao, tao-3., 3.-2.*tao, -tao},
-      std::array<float,4>{-tao, 2.-tao, tao-2., tao}
+  explicit CatmullRomSpline(const std::array<Point<dimensions>, N>& p, const double tao) : p(p), tao(tao) {
+    const double interval = (1.)/n_points;
+    std::array<std::array<double, 4>,  n_points+1> coefficients;
+    const std::array<std::array<double,4>, 4> tao_matrix {
+      std::array<double,4>{0., 1., 0., 0.},
+      std::array<double,4>{-tao, 0., tao, 0.},
+      std::array<double,4>{2.*tao, tao-3., 3.-2.*tao, -tao},
+      std::array<double,4>{-tao, 2.-tao, tao-2., tao}
     };
-    float u = 0;
+    double u = 0;
     for (int i = 0; i < n_points + 1; i++) {
-      const std::array<float, 4> u_matrix {1, u, u*u, u*u*u};
+      const std::array<double, 4> u_matrix {1, u, u*u, u*u*u};
       for (int j = 0; j < 4; j++) {
         coefficients[i][j] = 0;
         for (int k = 0; k < 4;  k++) {
@@ -56,7 +56,7 @@ public:
 private:
   std::array<Point<dimensions>, N> p;
   std::array<Point<dimensions>, (N-3)*(n_points+1)> points;
-  float tao;
+  double tao;
 };
 
 #endif  // CATMULL_ROM_SPLINE_HPP
